Add files via upload
Use these files first! Re the nemabiome-refined file refer to the R files for what sheets in the excel files to use.

For the R files:
Bioinformatic_pooling_new we pooled samples based on replicates, no pooling is when all three replicates are added up with no editing or removing of numbers. Bioinformatic pooling of two replicates is when data is added up. Samples that only had the species in one replicate are removed/ converted to zero. Bioinformatic pooling three replicates is when data is added up and the data is removed/ converted to zero if the species is not present in all 3 replicates. Ass three treatments were compared to the lab pooling of PCR reactions.